# Conflicts:
#	app/src/main/java/com/duckduckgo/app/browser/BrowserActivity.kt
#	app/src/main/java/com/duckduckgo/app/browser/BrowserTabFragment.kt
#	app/src/main/java/com/duckduckgo/app/browser/webview/BrowserContainerLayoutBehavior.kt

+++ pir/pir-impl/src/main/java/com/duckduckgo/pir/impl/brokers/BrokerDataDownloader.kt
@@ -20,21 +20,17 @@ import android.content.Context
 import com.duckduckgo.common.utils.DispatcherProvider
 import com.duckduckgo.di.scopes.AppScope
 import com.duckduckgo.pir.impl.service.DbpService
-import com.duckduckgo.pir.impl.service.DbpService.PirJsonBroker
-import com.duckduckgo.pir.impl.store.PirRepository
 import com.squareup.anvil.annotations.ContributesBinding
-import com.squareup.moshi.Moshi
-import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
+import kotlinx.coroutines.withContext
+import logcat.LogPriority.ERROR
+import logcat.logcat
+import okio.FileSystem.Companion.SYSTEM
+import okio.Path.Companion.toPath
+import okio.source
 import java.io.File
 import java.io.FileOutputStream
 import java.util.zip.ZipInputStream
 import javax.inject.Inject
-import kotlinx.coroutines.withContext
-import logcat.LogPriority.ERROR
-import logcat.logcat
-import okhttp3.ResponseBody
-import okio.FileSystem.Companion.SYSTEM
-import okio.Path.Companion.toPath
 
 interface BrokerDataDownloader {
     /**
@@ -50,38 +46,56 @@ interface BrokerDataDownloader {
 class RealBrokerDataDownloader @Inject constructor(
     private val dbpService: DbpService,
     private val dispatcherProvider: DispatcherProvider,
-    private val pirRepository: PirRepository,
     private val context: Context,
+    private val brokerJsonProcessor: BrokerJsonProcessor,
 ) : BrokerDataDownloader {
+
     override suspend fun downloadBrokerData(brokersToUpdate: List<String>) {
         withContext(dispatcherProvider.io()) {
             if (brokersToUpdate.isNotEmpty()) {
+                logcat { "PIR-update: Starting to download broker data" }
                 val extractFolder = File(context.filesDir, "unzipped-broker-json")
-                extractJsonFilesFromResponse(dbpService.getBrokerJsonFiles(), extractFolder)
-                processBrokerJsonFiles(extractFolder, brokersToUpdate)
-                SYSTEM.deleteRecursively(extractFolder.path.toPath())
+                val zipFile = File.createTempFile("broker-data", ".zip", context.cacheDir)
+                try {
+                    dbpService.getBrokerJsonFiles().use { body ->
+                        body.byteStream().use { input ->
+                            zipFile.outputStream().use { input.copyTo(it) }
+                        }
+                    }
+                    extractJsonFilesFromZip(zipFile, extractFolder)
+                    processBrokerJsonFiles(extractFolder, brokersToUpdate)
+                } finally {
+                    zipFile.delete()
+                    if (extractFolder.exists()) {
+                        SYSTEM.deleteRecursively(extractFolder.path.toPath())
+                    }
+                }
+                logcat { "PIR-update: Done downloading broker data" }
             }
         }
     }
 
-    private fun extractJsonFilesFromResponse(
-        responseBody: ResponseBody,
+    private fun extractJsonFilesFromZip(
+        zipFile: File,
         outputDir: File,
     ) {
-        logcat { "PIR-update: Extracting data from $responseBody" }
-        ZipInputStream(responseBody.byteStream()).use { zipInputStream ->
+        logcat { "PIR-update: Extracting data from $zipFile" }
+        outputDir.mkdirs()
+        val outputDirCanonical = outputDir.canonicalPath
+        ZipInputStream(zipFile.inputStream().buffered()).use { zipInputStream ->
             var entry = zipInputStream.nextEntry
             while (entry != null) {
-                val outputFile = File(outputDir, entry.name)
-
-                if (entry.isDirectory) {
-                    outputFile.mkdirs()
-                } else {
-                    FileOutputStream(outputFile).use { outputStream ->
-                        zipInputStream.copyTo(outputStream)
+                if (!entry.isDirectory && entry.name.endsWith(".json")) {
+                    val outputFile = File(outputDir, entry.name)
+                    if (outputFile.canonicalPath.startsWith(outputDirCanonical + File.separator)) {
+                        outputFile.parentFile?.mkdirs()
+                        FileOutputStream(outputFile).buffered().use { outputStream ->
+                            zipInputStream.copyTo(outputStream)
+                        }
+                    } else {
+                        logcat(ERROR) { "PIR-update: Skipping ZIP entry with invalid path: ${entry.name}" }
                     }
                 }
-
                 zipInputStream.closeEntry()
                 entry = zipInputStream.nextEntry
             }
@@ -99,21 +113,12 @@ class RealBrokerDataDownloader @Inject constructor(
             return
         }
 
-        val adapter = Moshi.Builder()
-            .add(KotlinJsonAdapterFactory())
-            .add(StepsAsStringAdapter())
-            .build()
-            .adapter(PirJsonBroker::class.java)
-
-        directory.listFiles()?.get(0)?.listFiles { file ->
-            file.extension == "json" && brokersToUpdate.contains(file.name)
+        val brokersToUpdateSet = brokersToUpdate.toHashSet()
+        directory.listFiles()?.firstOrNull()?.listFiles { file ->
+            file.extension == "json" && file.name in brokersToUpdateSet
         }?.forEach { jsonFile ->
             logcat { "PIR-update: Processing data from ${jsonFile.name}" }
-            val content = jsonFile.readText() // Read JSON file as string
-            val broker = adapter.fromJson(content)
-            if (broker != null) {
-                pirRepository.updateBrokerData(jsonFile.name, broker)
-            }
+            brokerJsonProcessor.processAndStoreBroker(jsonFile.name, jsonFile.source())
         }
 
         logcat { "PIR-update: Stored all new broker data" }
